improved coding style
