+++ src/main/java/frc/robot/subsystems/SwerveModule.java
@@ -14,7 +14,7 @@ import com.ctre.phoenix6.signals.NeutralModeValue;
 
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.PID;
+import frc.robot.util.PID;
 
 public class SwerveModule extends SubsystemBase {
   /** Creates a new SwerveModule. */
