+++ src/main/java/frc/robot/util/Util.java
@@ -1,0 +1,13 @@
+package frc.robot.util;
+
+import edu.wpi.first.wpilibj2.command.Command;
+import edu.wpi.first.wpilibj2.command.SubsystemBase;
+
+public final class Util {
+
+    public static <T extends SubsystemBase> T require(Command self, T requirement) {
+        self.addRequirements(requirement);
+        return requirement;
+    }
+
+}
